nt36xxx: Enable Low Power Wake-Up Gesture mode
As per LG Display Co Ltd

The touch display panel includes a source multiplexer (130) switching the transfer
of source signals to data lines (DL). In a low power wake-up gesture (LPWG) mode
in which a touch input is only sensed in a state in which a display function is off
to reduce power consumption, a source multiplexer control signal is provided to set
the source multiplexer (130) in an on state, in synchronization with a touch sensing
section in which a touch driving signal is provided. This can discharge electric
charges generated within the display panel, thereby preventing display defects due to
the electric charges.

Application number: 17188102.2
Publication date: 07.03.2018 Bulletin 2018/10

Signed-off-by: Pierre2324 <pdbbasketball@gmail.com>
Signed-off-by: CloudedQuartz <ravenklawasd@gmail.com>
Signed-off-by: jaaat4u <devangsingh369@gmail.com>