kernel: fair: turn off EAS_USE_NEED_IDLE sched feature
This feature disables the standard EAS task placement behavior for
prefer_idle tasks, causing them to prefer smaller idle CPUs instead of
the first idle CPU they find. For now let's use the  well-understood
EAS behavior we've used on previous devices, and revisit turning this
feature on if we decide it's an effective way to save power.

Signed-off-by: jaaat4u <devangsingh369@gmail.com>